[engine] [compile] update node struct (#4)
* [util] [test] TestRandomExpressions printProgress

* [engine] update flag type

* [engine] move child count and child index to

* [engine] make extra info slices

+++ parse_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 type verifyNode struct {
-	tpy       uint16
+	tpy       uint8
 	data      Value
 	cost      int
 	selectKey SelectorKey
